packages/ocrbridge-ocrmac/pyproject.toml
[project]
name = "ocrbridge-ocrmac"
version = "0.1.0"
description = "ocrmac (Apple Vision) OCR engine for OCR Bridge"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "ocrbridge-core>=0.1.0",
    "ocrmac>=0.2.2",
    "pdf2image>=1.17.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "pyright>=1.1.0",
]

# Entry point for engine discovery
[project.entry-points."ocrbridge.engines"]
ocrmac = "ocrbridge.engines.ocrmac:OcrmacEngine"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ocrbridge"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.11"
